refactor(package) : 도메인 패지키 분리
- Result , Reward , Rewards Domain reward 패키지로 이동
- Name Domain common 패키지로 이동
