+++ src/main/java/controller/GameController.java
@@ -3,10 +3,12 @@ package controller;
 import domain.*;
 import domain.ladder.Ladder;
 import domain.ladder.attribute.Height;
-import domain.player.Name;
+import domain.common.Name;
 import domain.player.Names;
 import domain.player.Player;
 import domain.player.Players;
+import domain.reward.Result;
+import domain.reward.Rewards;
 import util.RandomDirectionGenerator;
 import view.InputView;
 import view.OutputView;
