+++ src/main/java/domain/common/Name.java
@@ -1,4 +1,4 @@
-package domain.player;
+package domain.common;
 
 import java.util.Objects;
