+++ src/main/java/domain/player/Names.java
@@ -1,5 +1,7 @@
 package domain.player;
 
+import domain.common.Name;
+
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
