+++ src/main/java/domain/reward/Rewards.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.reward;
 
 
 import java.util.List;
