+++ src/test/java/domain/RewardTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.reward.Reward;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
