+++ src/test/java/domain/player/NameTest.java
@@ -1,6 +1,7 @@
 package domain.player;
 
 
+import domain.common.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
