+++ src/test/java/domain/player/PlayerTest.java
@@ -2,6 +2,7 @@ package domain.player;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import domain.common.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
